# docs/static/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: AltMount
    url: https://github.com/javi11/altmount/issues
  description: REST API for AltMount — Usenet WebDAV server with NZB queue, health
    monitoring, ARR integration, FUSE/rclone mounting, and Stremio addon.
  license:
    name: MIT
  termsOfService: http://altmount.kipsilabs.top
  title: AltMount API
  version: "1.0"
paths:
...
  "/config/providers/{id}/speedtest":
    post:
      description: Runs a speed test against the specified NNTP provider and saves the
        result to config.
      parameters:
        - description: Provider ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.ProviderSpeedTestResponse"
                    type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Test provider download speed
      tags:
        - Providers
...
servers:
  - url: //localhost:8080/api
components:
...
  securitySchemes:
    ApiKeyAuth:
      in: query
      name: apikey
      type: apiKey
  schemas:
    api.APIError:
      properties:
        code:
          type: string
        details:
          type: string
        message:
          type: string
      type: object
    api.APIMeta:
      properties:
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      type: object
    api.APIResponse:
      properties:
        data: {}
        error:
          $ref: "#/components/schemas/api.APIError"
        meta:
          $ref: "#/components/schemas/api.APIMeta"
        success:
          type: boolean
      type: object
...
    api.ProviderSpeedTestResponse:
      properties:
        duration_seconds:
          type: number
        speed_mbps:
          type: number
      type: object
...